requirements.txt
PySide6==6.3.1
feedparser==6.0.10
newspaper3k==0.2.8
requests==2.28.1
appdirs==1.4.4